Refactor Dialog Actions
