+++ src/components/CardFeed.js
@@ -2,5 +2,12 @@
  * Takes an array of cards as its child, and produces a scrollable feed
  */
 import React from 'react'
+import Grid from 'material-ui/Grid'
 
- export default (props) => (<div></div>)
+const CardFeed = (props) => (
+    <Grid container wrap='nowrap' direction={'column'} spacing={16}>
+        {props.children}
+    </Grid>
+)
+
+export default CardFeed
