+++ src/components/ClaimCard/CardActions/ArbiterActions.js
@@ -3,9 +3,9 @@
  * Defines the actions that an arbiter can perform on a claim
  */
 
-import CLAIM_STATES from './../ClaimStates'
-import {RuleOnClaim} from './../Actions'
-import {GenerateActionsFromMap} from './../util'
+import CLAIM_STATES from './ClaimStates'
+import RuleOnClaim from './../../DialogForms/RuleOnClaim'
+import {GenerateActionsFromMap} from './util'
 
 const ArbiterActionsByState = (claimState) => {
     switch(claimState){
