--- src/components/ClaimCard/CardActions/Validations.js
@@ -1,15 +1,0 @@
-const validateIncreaseClaimFee  = ({}) => {
-
-}
-const validateProposeSettlement = ({}) => {
-
-}
-const validateAcceptSettlement = ({}) => {
-
-}
-const validateSettlementFailed = ({}) => {
-
-}
-const validateRuleOnClaim = ({})  => {
-
-}
