+++ src/components/DialogForms/IncreaseClaimFee/validate.js
@@ -1,0 +1,6 @@
+import { validInteger, ValidateRequiredPositiveNumber } from "../validation";
+
+export default ({claimId, amount}) => ({
+    claimId:validInteger(claimId).error,
+    amount:ValidateRequiredPositiveNumber(amount).error
+})
