+++ src/components/DialogForms/ProposeSettlement/validate.js
@@ -1,0 +1,7 @@
+import { validInteger, ValidateRequiredPositiveNumber } from "../validation";
+
+export default ({claimId, amount}) => ({
+    claimId:validInteger(claimId).error,
+    amount:ValidateRequiredPositiveNumber(amount).error
+})
+
