+++ src/components/DialogForms/RuleOnClaim/validate.js
@@ -1,0 +1,1 @@
+export default ({claimId, ruling}) => ({})
