+++ src/components/StakeCard/CardActions/WhitelisteeActions.js
@@ -3,42 +3,10 @@
  * a stake card
  */
 
-import {openClaim} from '../../../services/delphi-contract/StakeActions'
-import {validateOpenClaim} from './WhitelisteeCardFormValidation'
+import OpenClaim from '../../DialogForms/OpenClaim';
 
-const WhitelisteeActions = (ethAddress, stakeAddress) => [
-    {
-        label:'Open a Claim',
-        dialogProps:{
-            title:'Open a Claim',
-            description:"Open a claim against this stake",
-            formName:'OpenClaimForm',
-            onSubmit:openClaim(ethAddress, stakeAddress),
-            validate:validateOpenClaim,
-            fields:[
-            {
-                type:'number',
-                label:'Amount',
-                name:'claimAmount',
-            },
-            {
-                type:'number',
-                label:'Fee',
-                name:'claimFee',
-            },
-            {
-                type:'text',
-                label:'Data',
-                name:'claimData'
-            },
-            {
-                type:'checkbox',
-                label:'Skip Settlement',
-                name:'claimSkipSettlement'
-            },
-            ],
-        }
-    }
-]
+const ApplyEthAddresses = (actions) => (ethAddress, stakeAddress) => actions.map(action => action(ethAddress, stakeAddress))
 
-export default WhitelisteeActions
+const WhitelisteeActions = [OpenClaim]
+
+export default ApplyEthAddresses(WhitelisteeActions)
