+++ src/pages/dashboard/index.js
@@ -62,7 +62,6 @@ const styles = theme => ({
       backgroundColor: theme.palette.background.default,
       padding: theme.spacing.unit * 3,
       minWidth: 0, // So the Typography noWrap works
-      backgroundColor:'#659dbd',
       backgroundImage: `url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='100' height='100' viewBox='0 0 200 200'%3E%3Cdefs%3E%3ClinearGradient id='a' gradientUnits='userSpaceOnUse' x1='100' y1='33' x2='100' y2='-3'%3E%3Cstop offset='0' stop-color='%23000' stop-opacity='0'/%3E%3Cstop offset='1' stop-color='%23000' stop-opacity='1'/%3E%3C/linearGradient%3E%3ClinearGradient id='b' gradientUnits='userSpaceOnUse' x1='100' y1='135' x2='100' y2='97'%3E%3Cstop offset='0' stop-color='%23000' stop-opacity='0'/%3E%3Cstop offset='1' stop-color='%23000' stop-opacity='1'/%3E%3C/linearGradient%3E%3C/defs%3E%3Cg fill='%235685a1' fill-opacity='0.06'%3E%3Crect x='100' width='100' height='100'/%3E%3Crect y='100' width='100' height='100'/%3E%3C/g%3E%3Cg fill-opacity='0.06'%3E%3Cpolygon fill='url(%23a)' points='100 30 0 0 200 0'/%3E%3Cpolygon fill='url(%23b)' points='100 100 0 130 0 100 200 100 200 130'/%3E%3C/g%3E%3C/svg%3E")`,
     },
     toolbar: theme.mixins.toolbar,
@@ -104,15 +103,13 @@ const ClippedDrawer = (props) => {
           <div className={classes.toolbar} />
           <List>
               {/* Drawer Buttons */}
-              {dashboardRoutes.map((route) => <SidebarLink path={route.path} name={route.name} icon={route.icon}/>)}
+              {dashboardRoutes.map((route, key) => <SidebarLink key={key} path={route.path} name={route.name} icon={route.icon}/>)}
           </List>
         </Drawer>
         <main className={classes.content}>
           <div className={classes.toolbar} />
           {/* Adds the switch router*/}
-          
               <SwitchRoutes location={location} userEthAddress={userEthAddress}/>
-            
         </main>
       </div>
     );
